+++ test/pdo_sqlsrv/pdo_278_lastinsertid_seq_2.phpt
@@ -7,52 +7,58 @@ include 'pdo_tools.inc';
 require_once("autonomous_setup.php");
 
 try{
-    $database = "master";
+    $database = "tempdb";
     $conn = new PDO("sqlsrv:Server=$serverName;Database=$database", $username, $password);
-
-    $tableName = GetTempTableName('tab');
-    $sequence1 = 'sequence1';
-    $sequence2 = 'sequenceNeg1';
-
-    $stmt = $conn->query("IF OBJECT_ID('$sequence1', 'SO') IS NOT NULL DROP SEQUENCE $sequence1");
-    $stmt = $conn->query("IF OBJECT_ID('$sequence2', 'SO') IS NOT NULL DROP SEQUENCE $sequence2");
-
-    $sql = "CREATE TABLE $tableName (ID INT IDENTITY(1,1), SeqNumInc INTEGER NOT NULL PRIMARY KEY, SomeNumber INT)";
-    $stmt = $conn->query($sql);
-
-    $sql = "CREATE SEQUENCE $sequence1 AS INTEGER START WITH 1 INCREMENT BY 1 MINVALUE 1 MAXVALUE 100";
-    $stmt = $conn->query($sql);
     
-    $sql = "CREATE SEQUENCE $sequence2 AS INTEGER START WITH 200 INCREMENT BY -1 MINVALUE 101 MAXVALUE 200";
-    $stmt = $conn->query($sql);
+    // sequence is only supported in SQL server 2012 and up (or version 11 and up)
+    // Output Done once the server version is found to be < 11
+    $version_arr = explode(".", $conn->getAttribute(PDO::ATTR_SERVER_VERSION));
+    if ($version_arr[0] < 11) {
+        echo "Done\n";
+    }
+    else {
+        $tableName = GetTempTableName('tab');
+        $sequence1 = 'sequence1';
+        $sequence2 = 'sequenceNeg1';
 
-    $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence1, 20 )");
-    $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence2, 180 )");
-    $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence1, 40 )");
-    $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence2, 160 )");
-    $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence1, 60 )");
-    $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence2, 140 )");
+        $stmt = $conn->query("IF OBJECT_ID('$sequence1', 'SO') IS NOT NULL DROP SEQUENCE $sequence1");
+        $stmt = $conn->query("IF OBJECT_ID('$sequence2', 'SO') IS NOT NULL DROP SEQUENCE $sequence2");
 
-    // return the last sequence number of 'sequence1'
-    $lastSeq = $conn->lastInsertId($sequence1);
-    echo ("Last Sequence:\n");
-    var_dump($lastSeq);
+        $sql = "CREATE TABLE $tableName (ID INT IDENTITY(1,1), SeqNumInc INTEGER NOT NULL PRIMARY KEY, SomeNumber INT)";
+        $stmt = $conn->query($sql);
+
+        $sql = "CREATE SEQUENCE $sequence1 AS INTEGER START WITH 1 INCREMENT BY 1 MINVALUE 1 MAXVALUE 100";
+        $stmt = $conn->query($sql);
     
-    // return the last sequence number of 'sequenceNeg1'
-    $lastSeq = $conn->lastInsertId($sequence2);
-    echo ("\nLast Sequence:\n");
-    var_dump($lastSeq);
+        $sql = "CREATE SEQUENCE $sequence2 AS INTEGER START WITH 200 INCREMENT BY -1 MINVALUE 101 MAXVALUE 200";
+        $stmt = $conn->query($sql);
+
+        $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence1, 20 )");
+        $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence2, 180 )");
+        $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence1, 40 )");
+        $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence2, 160 )");
+        $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence1, 60 )");
+        $ret = $conn->exec("INSERT INTO $tableName VALUES( NEXT VALUE FOR $sequence2, 140 )");
+
+        // return the last sequence number of 'sequence1'
+        $lastSeq1 = $conn->lastInsertId($sequence1);
     
-    // providing a table name in lastInsertId should return an empty string
-    $lastSeq = $conn->lastInsertId($tableName);
-    echo ("\nLast Sequence:\n");
-    var_dump($lastSeq);
+        // return the last sequence number of 'sequenceNeg1'
+        $lastSeq2 = $conn->lastInsertId($sequence2);
+    
+        // providing a table name in lastInsertId should return an empty string
+        $lastSeq3 = $conn->lastInsertId($tableName);
+        
+        if ($lastSeq1 == 3 && $lastSeq2 == 198 && $lastSeq3 == "") {
+            echo "Done\n";
+        }
 
-    $stmt = $conn->query("DROP TABLE $tableName");
-    $stmt = $conn->query("DROP SEQUENCE $sequence1");
-    $stmt = $conn->query("DROP SEQUENCE $sequence2");
+        $stmt = $conn->query("DROP TABLE $tableName");
+        $stmt = $conn->query("DROP SEQUENCE $sequence1");
+        $stmt = $conn->query("DROP SEQUENCE $sequence2");
 
-    $stmt = null;
+        $stmt = null;
+    }
     $conn = null;
 }
     catch (Exception $e){
@@ -61,11 +67,4 @@ try{
    
 ?>
 --EXPECT--
-Last Sequence:
-string(1) "3"
-
-Last Sequence:
-string(3) "198"
-
-Last Sequence:
-string(0) ""
+Done
